+++ R/integration.R
@@ -321,7 +321,24 @@ fm_int.list <- function(domain, samplers = NULL, ...) {
   index_single_samplers <- which(names_lsamplers != "")
   index_multi_samplers <- which(names_lsamplers == "")
   names_samplers <- as.list(names_lsamplers)
-  names_samplers[index_multi_samplers] <- lapply(samplers[index_multi_samplers], names)
+  names_samplers[index_multi_samplers] <-
+    lapply(
+      samplers[index_multi_samplers],
+      function(x) {
+        if (inherits(x, "fm_segm")) {
+          stop(
+            paste0(
+              "Unnamed sampler in the samplers is an 'fm_segm' object.\n",
+              "Use an 'sf' object or other supported multi-sampler class instead."
+            ),
+            immediate. = TRUE
+          )
+          NULL
+        } else {
+          names(x)
+        }
+      }
+    )
   names_reserved <- c("weight", ".block") # coordinate and geometry is not required here
 
   if (length(intersect(names_domain, names_reserved)) > 0) {
@@ -339,7 +356,7 @@ fm_int.list <- function(domain, samplers = NULL, ...) {
   # TODO still have to deal with secondary geometry
   for (i in index_multi_samplers) {
     if (is.null(names_samplers[[i]])) {
-      stop(paste0("The unnamed sampler #", i, " in the samplers is NULL"))
+      stop(paste0("The unnamed sampler #", i, " in the samplers has no sub-names."))
     }
     lips_samplers[[i]] <-
       fm_int_multi_sampler(
@@ -353,7 +370,9 @@ fm_int.list <- function(domain, samplers = NULL, ...) {
   # singledomain samplers, ie named element(s) in samplers
   for (i in index_single_samplers) {
     nm <- intersect(names_samplers[[i]], names_domain)
-    stopifnot(length(nm) == 1)
+    if (length(nm) == 0) {
+      stop(paste0("The named sampler '", names_lsamplers[[i]], "' in the samplers has no corresponding domain."))
+    }
     lips_samplers[[i]] <-
       fm_int(
         domain = domain[[nm]],